docker/data_ingester/requirements.txt
requests==2.29.0
pandas==1.5.3
pyspark==3.4.1
google-cloud-storage==2.10.0